Moves the manifest controller testing helpers to the root testing package
