+++ tests/integration/controller/manifest/custom_resource_check/suite_test.go
@@ -44,7 +44,7 @@ import (
 	"github.com/kyma-project/lifecycle-manager/internal/manifest"
 	"github.com/kyma-project/lifecycle-manager/pkg/log"
 	"github.com/kyma-project/lifecycle-manager/tests/integration"
-	"github.com/kyma-project/lifecycle-manager/tests/integration/controller/manifest/manifesttest"
+	manifesttest "github.com/kyma-project/lifecycle-manager/tests/integration/controller/manifest"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
